feat(core): cleans up hashing logic
