+++ Core/Core/Logging/Analytics.cs
@@ -85,8 +85,8 @@ namespace Speckle.Core.Logging
         if (acc == null)
           return;
 
-        email = acc.userInfo.email;
-        server = acc.serverInfo.url;
+        email = acc.GetHashedEmail();
+        server = acc.GetHashedServer();
       }
 
       TrackEvent(email, server, eventName, customProperties);
@@ -100,23 +100,20 @@ namespace Speckle.Core.Logging
     /// <param name="customProperties">Additional parameters to pass to the event</param>
     public static void TrackEvent(Account account, Events eventName, Dictionary<string, object> customProperties = null)
     {
-      string email = account?.userInfo?.email ?? "unknown";
-      string url = account?.serverInfo?.url ?? "https://speckle.xyz/";
-
-      TrackEvent(email, url, eventName, customProperties);
+      TrackEvent(account.GetHashedEmail(), account.GetHashedServer(), eventName, customProperties);
     }
 
     /// <summary>
     /// Tracks an event from a specified email and server, anonymizes personal information
     /// </summary>
-    /// <param name="email">Email of the user, it will be anonymized</param>
-    /// <param name="server">Server URL, it will be anonymized</param>
+    /// <param name="hashedEmail">Email of the user anonymized</param>
+    /// <param name="hashedServer">Server URL anonymized</param>
     /// <param name="eventName">Name of the event</param>
     /// <param name="customProperties">Additional parameters to pass to the event</param>
-    private static void TrackEvent(string email, string server, Events eventName, Dictionary<string, object> customProperties = null)
+    private static void TrackEvent(string hashedEmail, string hashedServer, Events eventName, Dictionary<string, object> customProperties = null)
     {
-      LastEmail = email;
-      LastServer = server;
+      LastEmail = hashedEmail;
+      LastServer = hashedServer;
 
 #if DEBUG
       //only track in prod
@@ -128,10 +125,6 @@ namespace Speckle.Core.Logging
 
         try
         {
-          server = CleanURL(server);
-          var hashedEmail = "@" + Hash(email); //prepending an `@` (not relevant anymore but let's keep it for consistency)
-          var hashedServer = Hash(server);
-
           var properties = new Dictionary<string, object>()
           {
             { "distinct_id", hashedEmail },
@@ -169,13 +162,12 @@ namespace Speckle.Core.Logging
 
     }
 
-    internal static void AddConnectorToProfile(string email, string connector)
+    internal static void AddConnectorToProfile(string hashedEmail, string connector)
     {
       Task.Run(() =>
       {
         try
         {
-          var hashedEmail = "@" + Hash(email);
           var data = new Dictionary<string, object>()
           {
             { "$token", MixpanelToken },
@@ -203,34 +195,9 @@ namespace Speckle.Core.Logging
       });
     }
 
-    private static string CleanURL(string server)
-    {
-      Uri NewUri;
 
-      if (Uri.TryCreate(server, UriKind.Absolute, out NewUri))
-      {
-        server = NewUri.Authority;
-      }
-      return server;
-    }
 
-    private static string Hash(string input)
-    {
 
-      using (System.Security.Cryptography.MD5 md5 = System.Security.Cryptography.MD5.Create())
-      {
-        byte[] inputBytes = System.Text.Encoding.ASCII.GetBytes(input.ToLowerInvariant());
-        byte[] hashBytes = md5.ComputeHash(inputBytes);
-
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < hashBytes.Length; i++)
-        {
-          sb.Append(hashBytes[i].ToString("X2"));
-        }
-        return sb.ToString();
-      }
-
-    }
 
     private static string GetOs()
     {
